The implementation is almost done but.
The implementation is almost done but the box.hpp
has a problem. It outputs weird normals it can be a problem
of rectangles thus aarect or others like
materials textures etc.
The hitrecord now holds an object type string
for facilitating debugging
There is also a dc color variable which
should be checked if its value is nan.
When its value is nan it means pdf_val and
scattering_pdf resulted in 0 at the same time
this happens mostly due to weird normals
that are feeded into cosinepdf of lambertian
material, I think.

Once the implementation done this is going
into annex of thesis in its integrity.